requirements.txt
python-telegram-bot==21.10
supabase==2.11.0
python-dotenv==1.0.1
websockets>=13.0
